Fix segmentation fault due to missing reference.
- The GrassFoliage objects were trying to retrieve a non existant FoliageLevelManager from Foliage to connect to. It should be created before the foliage is initialized.

- GrassFoliage was also trying to connect to a signal that sent int instead of float.

+++ src/components/ogre/environment/FoliageLevelManager.h
@@ -38,7 +38,7 @@ public:
 	/**
 	 * @brief Signal sent when this component changes foliage density levels.
 	 */
-	sigc::signal<void, int> foliageDensityChanged;
+	sigc::signal<void, float> foliageDensityChanged;
 
 	/**
 	 * @brief Used to change the density of all foliage.
